Fix Typo
I'm surprised this went unnoticed for so long; though when lookin through older projects I ran into it. I've contemplated making this font use the unicode format instead for ease of editing, but if I do that I'll probably include a full font image still (or something like a .psd that's the workspace I use).